style: format pull request template
