+++ .github/pull_request_template.md
@@ -1,7 +1,7 @@
 # UX Stage 0X: <short, imperative title>
 
 > **Plan adherence:** I read and followed [docs/ux-upgrade-plan.md] (plan version: **\_).
-> If not, explain why: **********\_\_\_\_************
+> If not, explain why: **\*\*\*\*\***\*\_\_\_\_\*\***\*\*\*\*\*\*\*\*
 
 ## Summary
